requirements.txt
python-dotenv
google-generativeai
langchain
langchain-community
pyPDF2
faiss-cpu
streamlit
-e .